# agents/src/agents/iri_api_spec.yaml
openapi: 3.0.3
info:
  title: IRI Annuity Renewal Intelligence API
  description: API specification for the IRI Dashboard backend services
  version: 1.1.0
  contact:
    name: IRI Team 2
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://api.iri-hackathon.com/api
    description: Production server
paths:
...
  /alerts:
    get:
      summary: Get all renewal alerts
      description: Retrieves all active renewal alerts for the advisor's portfolio (lightweight list data)
      operationId: getAlerts
      tags: [Dashboard]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, reviewed, snoozed, dismissed]
        - name: priority
          in: query
          schema:
            type: string
            enum: [high, medium, low]
        - name: carrier
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RenewalAlert'
...
  /alerts/{alertId}:
    get:
      summary: Get alert detail
      description: Returns full alert detail including policy data, suitability score, disclosures, transaction options, and audit log. Used when advisor clicks into an alert.
      operationId: getAlertDetail
      tags: [Alert Detail]
      parameters:
        - $ref: '#/components/parameters/AlertId'
      responses:
        '200':
          description: Full alert detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /alerts/{alertId}/snooze:
    post:
      summary: Snooze an alert
      operationId: snoozeAlert
      tags: [Alert Actions]
      parameters:
        - $ref: '#/components/parameters/AlertId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [snoozeDays]
              properties:
                snoozeDays:
                  type: integer
                  minimum: 1
                  maximum: 90
                reason:
                  type: string
      responses:
        '200':
          description: Alert snoozed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
                  snoozeUntil: { type: string, format: date-time }
  /alerts/{alertId}/dismiss:
    post:
      summary: Dismiss an alert
      operationId: dismissAlert
      tags: [Alert Actions]
      parameters:
        - $ref: '#/components/parameters/AlertId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
      responses:
        '200':
          description: Alert dismissed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }

  # ── Compare Tab ────────────────────────────────────────────
...
  /alerts/{alertId}/suitability:
    put:
      summary: Save suitability data
      description: Saves the suitability form data for this alert review
      operationId: saveSuitability
      tags: [Action]
      parameters:
        - $ref: '#/components/parameters/AlertId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SuitabilityData'
      responses:
        '200':
          description: Suitability saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
...
components:
...
  parameters:
    AlertId:
      name: alertId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  # ── Schemas ────────────────────────────────────────────────
  schemas:
    AlertType:
      type: string
      enum: [replacement_recommended, replacement_opportunity, missing_info, suitability_review, income_planning]
    RenewalAlert:
      type: object
      required: [id, clientId, policyId, clientName, carrier, renewalDate, daysUntilRenewal, currentRate, renewalRate, currentValue, isMinRate, priority, hasDataException, status, alertType, alertDescription]
      properties:
        id: { type: string, example: "alert-ANN-2020-5621-renewal" }
        clientId: { type: string, example: "CLI-1001" }
        policyId: { type: string, example: "ANN-2020-5621" }
        clientName: { type: string, example: "Maria Rodriguez" }
        carrier: { type: string, example: "Athene" }
        renewalDate: { type: string, example: "15 Days" }
        daysUntilRenewal: { type: integer, example: 15 }
        currentRate: { type: string, example: "3.8%" }
        renewalRate: { type: string, example: "1.5%" }
        currentValue: { type: string, example: "$180,000" }
        isMinRate: { type: boolean, example: true }
        priority: { type: string, enum: [high, medium, low] }
        hasDataException: { type: boolean }
        missingFields:
          type: array
          items: { type: string }
        status: { type: string, enum: [pending, reviewed, snoozed, dismissed] }
        alertType: { $ref: '#/components/schemas/AlertType' }
        alertTypes:
          type: array
          items: { $ref: '#/components/schemas/AlertType' }
        alertDescription: { type: string }
...
    AlertDetail:
      type: object
      required: [alert, clientAlerts, policyData, aiSuitabilityScore, suitabilityData, disclosureItems, transactionOptions, auditLog]
      properties:
        alert:
          $ref: '#/components/schemas/RenewalAlert'
        clientAlerts:
          type: array
          items: { $ref: '#/components/schemas/RenewalAlert' }
          description: All alerts for this client (for grouped review)
        policyData:
          $ref: '#/components/schemas/PolicyData'
        aiSuitabilityScore:
          $ref: '#/components/schemas/SuitabilityScore'
        suitabilityData:
          $ref: '#/components/schemas/SuitabilityData'
        disclosureItems:
          type: array
          items: { $ref: '#/components/schemas/DisclosureItem' }
        transactionOptions:
          type: array
          items: { $ref: '#/components/schemas/TransactionOption' }
        auditLog:
          type: array
          items: { $ref: '#/components/schemas/AuditLogEntry' }
    PolicyData:
      type: object
      required: [contractId, clientName, carrier, productName, issueDate, currentValue, surrenderValue, currentRate, renewalRate, guaranteedMinRate, renewalDate, isMinRateRenewal, suitabilityStatus, eligibilityStatus, features]
      properties:
        contractId: { type: string, example: "ANN-2020-5621" }
        clientName: { type: string }
        carrier: { type: string }
        productName: { type: string, example: "SecureChoice Multi-Year Guarantee Annuity" }
        issueDate: { type: string, example: "03/15/2018" }
        currentValue: { type: string, example: "$180,000" }
        surrenderValue: { type: string, example: "$176,400" }
        currentRate: { type: string, example: "3.8%" }
        renewalRate: { type: string, example: "1.5%" }
        guaranteedMinRate: { type: string, example: "1.50%" }
        renewalDate: { type: string }
        isMinRateRenewal: { type: boolean }
        suitabilityStatus: { type: string, enum: [complete, incomplete, not-started] }
        eligibilityStatus: { type: string, enum: [eligible, restricted, ineligible] }
        features:
          $ref: '#/components/schemas/PolicyFeaturesData'
    PolicyFeaturesData:
      type: object
      properties:
        features:
          type: array
          items: { $ref: '#/components/schemas/PolicyFeature' }
        benefits:
          type: array
          items: { $ref: '#/components/schemas/PolicyFeature' }
        riders:
          type: array
          items: { $ref: '#/components/schemas/PolicyFeature' }
        limitations:
          type: array
          items: { $ref: '#/components/schemas/PolicyFeature' }
        surrenderCharge: { type: string, example: "2.5% (Year 7 of 10)" }
        withdrawalAllowance: { type: string, example: "10% annually penalty-free" }
        mvaPenalty: { type: string, example: "-3.8% if surrendered today" }
        rateGuarantee: { type: string, example: "5 years remaining" }
    PolicyFeature:
      type: object
      required: [name, description, included, type]
      properties:
        name: { type: string }
        description: { type: string }
        included: { type: boolean }
        type: { type: string, enum: [feature, benefit, rider, limitation] }
    SuitabilityScore:
      type: object
      required: [score, reasoning, complianceNotes, missingData]
      properties:
        score: { type: integer, example: 85 }
        reasoning:
          type: array
          items: { type: string }
        complianceNotes:
          type: array
          items: { type: string }
        missingData:
          type: array
          items: { type: string }
    SuitabilityData:
      type: object
      required: [clientObjectives, riskTolerance, timeHorizon, liquidityNeeds, taxConsiderations, guaranteedIncome, rateExpectations, surrenderTimeline, livingBenefits, advisorEligibility, score, isPrefilled]
      properties:
        clientObjectives: { type: string }
        riskTolerance: { type: string }
        timeHorizon: { type: string }
        liquidityNeeds: { type: string }
        taxConsiderations: { type: string }
        guaranteedIncome: { type: string }
        rateExpectations: { type: string }
        surrenderTimeline: { type: string }
        livingBenefits:
          type: array
          items: { type: string }
        advisorEligibility: { type: string }
        score: { type: integer }
        isPrefilled: { type: boolean }
    DisclosureItem:
      type: object
      required: [id, required]
      properties:
        id: { type: string }
        label: { type: string }
        title: { type: string }
        description: { type: string }
        documentName: { type: string }
        required: { type: boolean }
        link: { type: string }
    TransactionOption:
      type: object
      required: [type, label, description, requirements, isAvailable]
      properties:
        type: { type: string, enum: [renew, replace] }
        label: { type: string }
        description: { type: string }
        requirements:
          type: array
          items: { type: string }
        warnings:
          type: array
          items: { type: string }
        isAvailable: { type: boolean }
        unavailableReason: { type: string }
        pros:
          type: array
          items: { type: string }
        cons:
          type: array
          items: { type: string }
    AuditLogEntry:
      type: object
      required: [timestamp, user, action, details]
      properties:
        timestamp: { type: string }
        user: { type: string }
        action: { type: string }
        details: { type: string }

    # ── Compare ──────────────────────────────────────────────
...
    Error:
      type: object
      required: [error, message]
      properties:
        error: { type: string, example: "NOT_FOUND" }
        message: { type: string, example: "Alert not found" }
        details:
          type: object
          additionalProperties: true